# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Mate API",
        "description": "API documentation for the Travel Mate project",
        "contact": {
            "email": "support@travelmate.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Lister tous les voyages",
                "operationId": "692b0f6b729af25639ef76046f8fba1f",
                "responses": {
                    "200": {
                        "description": "Liste de tous les voyages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "trips": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Voyage à Paris"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Une belle aventure en Europe"
                                                    },
                                                    "destination": {
                                                        "type": "string",
                                                        "example": "Europe"
                                                    },
                                                    "duration": {
                                                        "type": "integer",
                                                        "example": 7
                                                    },
                                                    "city_id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "activities": {
                                                        "type": "string",
                                                        "example": "Randonnée, Visite des musées"
                                                    },
                                                    "included": {
                                                        "type": "string",
                                                        "example": "Transport, Hébergement"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "https://example.com/image.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "Créer un nouveau voyage (admin seulement)",
                "operationId": "330c41008c88334c0e7653cbe4f0710a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "destination",
                                    "duration",
                                    "activities",
                                    "included",
                                    "city_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Voyage à Paris"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Une belle aventure en Europe."
                                    },
                                    "destination": {
                                        "type": "string",
                                        "enum": [
                                            "Europe",
                                            "Amérique",
                                            "Afrika",
                                            "Asie",
                                            "Australie"
                                        ],
                                        "example": "Europe"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "activities": {
                                        "type": "string",
                                        "example": "Randonnée, Visite des musées"
                                    },
                                    "included": {
                                        "type": "string",
                                        "example": "Transport, Hébergement"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voyage créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "trip": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Voyage à Paris"
                                                },
                                                "destination": {
                                                    "type": "string",
                                                    "example": "Europe"
                                                },
                                                "duration": {
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "activities": {
                                                    "type": "string",
                                                    "example": "Randonnée, Visite des musées"
                                                },
                                                "included": {
                                                    "type": "string",
                                                    "example": "Transport, Hébergement"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "https://example.com/image.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé. Vous devez être administrateur.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Accès refusé. Vous devez être administrateur."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}